% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleandata}
\alias{cleandata}
\title{Description of NPS data}
\format{a \code{data frame} 

\describe{
\item{this}{this}
\item{this}{that}
}}
\description{
A dataset imputing values for the original NPS data
}
