% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{getCI}
\alias{getCI}
\title{get CI score}
\usage{
getCI(groups, countsL, plot = TRUE, adjust.size = FALSE, ylim = NULL,
  nr = 1, nc = length(countsL), order = NULL)
}
\arguments{
\item{groups}{A list of vertex id vectors}

\item{countsL}{a list of numeric count matrix}

\item{plot}{a boolean to decide if plot a bar plot of CI scores or not. Default TRUE}

\item{adjust.size}{a boolean to decide if CI score should be adjust by size or not. Default FALSE}

\item{ylim}{a vector needed if the output barplots need to be on the same y scale}

\item{nr}{the number of rows to plot}

\item{nc}{the number of column to plot}

\item{order}{the order of the barplot. default NULL which is using the input list order}
}
\value{
a list of CI score and their components
}
\description{
get CI score
}
\examples{
par(mfrow <- c(1,5))
membersL_noweight <- getCI(cluster,test,adjust.size = F,ylim = c(0,8))
head(membersL_noweight)
maxCIms <- getMaxCImember(membersL_noweight[[1]],membersL_noweight[[2]],min =3)
head(maxCIms)

}
\author{
Zhezhen Wang
}
