% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{getCluster}
\alias{getCluster}
\title{Grouping Similar Datasets}
\usage{
getCluster(igraphL, steps = 4)
}
\arguments{
\item{igraphL}{A list of count matrix or a list of igrap objects}

\item{steps}{A number of steps}
}
\value{
A list of objects with similar correlations within networks.
}
\description{
Samples with relatively similar correlation scores are grouped
together. Several clusters with their respective correlation groups
could be present within a networked nodes.
}
\examples{
gc <- getCluster(igraphL, steps = 4)
gc
}
\author{
Zhezhen Wang
}
