% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{getCluster_methods}
\alias{getCluster_methods}
\title{get clusters of nodes by clustering methods}
\usage{
getCluster_methods(igraphL, method = "rw", cutoff = NULL)
}
\arguments{
\item{igraphL}{a list of count matrix or a list of igraph object}

\item{method}{a character need to be select from 'rw', 'hcm', 'km', 'pam', or
'natrual'. Uses 'rw' by default for 'hcm', using complete. for 'km' using
'euclidean'.}

\item{cutoff}{a numeric value, default NULL}

\item{countsL}{a list of count matrix}
}
\value{
A list of vertex id vectors that can be used in plot function in igraph package as 'mark.groups' parameter
}
\description{
get clusters of nodes by clustering methods
}
\examples{
cluster <- getCluster_methods(igraphL)
head(cluster)

}
\author{
Zhezhen Wang and Biniam Feleke
}
