% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_paracoords.R
\name{myparacoords}
\alias{myparacoords}
\title{Parallel Coordinate Plot}
\usage{
myparacoords(dataSource)
}
\arguments{
\item{dataSource}{A data.frame with a column called "LCLUCI.labels"}
}
\description{
uses timelyportfolio's github function for a parallel coordinate plots
}
\examples{
\dontrun{myparacoords(cleandata[, c("LCLUCI.labels","Developed200m",	 "RecCon5km",	 "DistCoast", "WaterOnly200m",	"DistHighAirports", "DistModerateAirports", 	"DistRoadsMajor",	"RddMajorPt", "park", "L90f1", "L90f3", "L90f5", "L90dBA")])}

}
