% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_package.R
\docType{package}
\name{NPS-package}
\alias{NPS}
\alias{NPS-package}
\title{Convenience Functions and Tests for Working With the Net Promoter Score (NPS)}
\description{
Small functions to make working with survey data in the context of a Net
Promoter programme easier. Specifically, data transformation methods, some
methods for examining the statistical properties of the NPS, such as its
variance and standard errors, and some simple inferential testing procedures.
Net Promoter and NPS are registered trademarks of Bain & Company, Satmetrix
Systems and Fred Reichheld.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
\keyword{package}

