% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nps.se}
\alias{nps.se}
\title{Calculate the standard error of a Net Promoter Score}
\usage{
nps.se(x, breaks = list(0:6, 7:8, 9:10))
}
\arguments{
\item{x}{A vector of \emph{Recommend} scores}

\item{breaks}{A \code{list} of length three, giving the integer Likert scale points for \emph{Detractors}, \emph{Passives}, and \emph{Promoters}, respectively. The default is \code{list(0:6, 7:8, 9:10)}}
}
\value{
\code{\link{numeric}}. The variance of the distribution, ranging from 0 to 1.
}
\description{
This function calculates the standard error (see below) of a Net Promoter Score, taking a \code{\link{vector}} of raw \emph{Recommend} data
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\seealso{
\code{\link{npvar}}, a version which works on counts or proportions of responses
}

