% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_global.R
\name{nps2_name_check}
\alias{nps2_name_check}
\title{Small function to emit warnings for using functions in the old syntax.}
\usage{
nps2_name_check()
}
\description{
Small function to emit warnings for using functions in the old syntax.
}

