% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_global.R
\name{nps_options}
\alias{nps_options}
\title{Display global options for NPS}
\usage{
nps_options()
}
\value{
A \code{list} of the options available
}
\description{
Run \code{nps_options()} to see what it's possible to set globally.
}
\details{
Options can be set in the usual way using \code{\link{options}}, and include:

\describe{
  \item{\code{nps.breaks}}{
    A \code{\link{numeric}} \code{\link{vector}} of length 4, containing the
    extent of the measurement scale used for calculating NPS, and the
    'breaks' for transforming it into categories.

    The four entries are:

    1. The lowest valid value of the scale

    2. The highest value for a Detractor

    3. The highest value for a Passive

    4. The highest valid value of the scale

    So, for example, the 'best practice' values for Net Promoter (and the
    package's default), is the vector \code{c(0, 6, 8, 10)}.
  }
  \item{\code{nps.100}}{
    By default, \code{NPS} presents scores as being between -1 and 1.
    However, scores are often multiplied by 100 for presentational purposes
    (producing a score between -100 and 100). Should the \code{NPS} package
    do this by default? \code{\link{logical}}.
  }
}
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}

