% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_tests.R
\name{nps_test}
\alias{nps_test}
\alias{nps_test_}
\alias{print.nps_test}
\alias{print.nps_test,}
\title{Significance tests and confidence intervals for Net Promoter Scores}
\usage{
nps_test(x, y = NULL, test = "wald", conf = 0.95,
  breaks = getOption("nps.breaks"))

nps_test_(x, y = NULL, test = "wald", conf = 0.95)

\method{print}{nps_test}(x, ...)
}
\arguments{
\item{x}{A vector of \emph{Recommend} scores}

\item{y}{A vector of \emph{Recommend} scores, to compare to \code{x}. If not
specified, a one sample test on \code{x} is performed. Defaults to
\code{NULL}}

\item{test}{The type of test to perform. Currently only the Wald/Z-test
('\code{wald}') is supported}

\item{conf}{the confidence level of the test and intervals. Defaults to 0.95}

\item{breaks}{A \code{vector} of length three, giving integer Likert scale
break-points for Net Promoter categories. Best practice is to set the
values you'd like to use at the start of a session, with
\code{options(nps.breaks = c(0, 6, 8, 10))} (for the usual setting of 0-6
being \emph{Detractors}, 7-8 being \emph{Passives} and 9-10 being
\emph{Promoters}, also the package default), or whichever values you would
like. However, a vector may be supplied directly.}

\item{\dots}{Not used.}
}
\value{
A \code{\link{list}} of class \code{nps.test} containing:
  \item{nps.x, nps.y}{The Net Promoter score(s)} \item{delta}{Where both
  \code{x} and \code{y} have been specified, the absolute difference between
  the two scores} \item{int}{The confidence interval generated. For a one
  sample test, this will be a confidence interval around \code{nps.x}. For a
  two sample test, this will be a confidence interval around the difference
  between \code{nps.x} and \code{nps.y}} \item{conf}{The confidence level
  used when performing the test, as specificed by \code{conf} in the function
  parameters} \item{p.value}{The p value of the significance test}
  \item{sig}{\code{\link{logical}}. Whether or not the \code{p.value} of the
  test exceeded 1-\code{conf}} \item{se.hat}{The estimated standard error of
  \code{delta} for a two sample test, otherwise of \code{x}}
  \item{type}{\code{\link{character}} string indicating whether a one or two
  sample test was performed} \item{n.x, n.y}{Counts for \code{x} and
  \code{y}}

\code{NULL}
}
\description{
This function performs one and two sample tests for the Net Promoter score(s)
of \emph{Recommend} data distributions. Currently, only a Wald type test is
supported.
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
\seealso{
\code{\link{nps_var}}, \code{\link{nps_se}}, \code{\link{nps}}
}

