% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps.R
\name{nps_var}
\alias{nps.var}
\alias{nps_var}
\alias{nps_var_}
\alias{npvar}
\title{Calculate the variance of a Net Promoter Score}
\usage{
nps_var(x, breaks = getOption("nps.breaks"), na.rm = FALSE)

nps_var_(x, na.rm = FALSE)

nps.var(...)

npvar(...)
}
\arguments{
\item{x}{A vector of \emph{Recommend} scores}

\item{breaks}{A \code{vector} of length three, giving integer Likert scale
break-points for Net Promoter categories. Best practice is to set the
values you'd like to use at the start of a session, with
\code{options(nps.breaks = c(0, 6, 8, 10))} (for the usual setting of 0-6
being \emph{Detractors}, 7-8 being \emph{Passives} and 9-10 being
\emph{Promoters}, also the package default), or whichever values you would
like. However, a vector may be supplied directly.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
\code{\link{numeric}}. The variance of the distribution, ranging from
  0 to 1.
}
\description{
This function calculates the Net Promoter Score variance, taking a
\code{\link{vector}} of raw \emph{Recommend} data
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
\seealso{
\code{\link{nps_var_}}, a version which works on counts or
  proportions of responses
}

