% Generated by roxygen2 (4.0.2): do not edit by hand
\name{npvar}
\alias{npvar}
\title{Calculate the variance of a Net Promoter Score}
\usage{
npvar(x)
}
\arguments{
\item{x}{A \code{\link{vector}} of length three, with numbers or proportions of \emph{Detractors}, \emph{Passives}, and \emph{Promoters}, respectively}
}
\value{
\code{\link{numeric}}. The variance of the distribution, ranging from 0 to 1.
}
\description{
This function calculates the Net Promoter Score variance, taking a \code{\link{vector}} of length three, with numbers or proportions of \emph{Detractors}, \emph{Passives}, and \emph{Promoters}, respectively.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\seealso{
\code{\link{nps.var}}, a version which works on raw Recommend responses
}

