% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_my_app.R
\name{run_my_app}
\alias{run_my_app}
\title{Runs an app in my package}
\usage{
run_my_app(app_name, ...)
}
\arguments{
\item{app_name}{character string for a directory in this package}

\item{...}{additional options passed to shinyAppdir}
}
\value{
a printed shiny app
}
\description{
Runs as shiny app
}
\examples{
#remember to use EXAMPLES not Example
\dontrun{
run_my_app("myFirstApp")
}
}
\author{
Rachel
}
