% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{sd_selection}
\alias{sd_selection}
\title{The length of a string (in characters).}
\usage{
sd_selection(df, samplesL, cutoff = 0.01, method = "other")
}
\arguments{
\item{df}{a count matrix with unique loci row names:ID X column names:samples.}

\item{samplesL}{a list of characters with stages as names.}

\item{cutoff}{numeric value, if < 1 automaticlly goes to select top x% transcripts
if > 1 goes to select by x-fold more than the selected method (which is
either the reference, other stages or pervious stage) default 0.01.}

\item{method}{select from 'reference','other', or 'previous'.
for 'reference', the reference has to be the first
for 'previous', make sure sampleL is in the right order from benign to malign
default uses 'other'}
}
\value{
numeric vector giving number of characters in each element of the
character vector. Missing strings have missing length.
}
\description{
The length of a string (in characters).
}
\examples{
GSE6136 = read.table("C:/Users/benjk/Desktop/GSE6136_matrix.txt", header = TRUE, comment = '!')
row.names(GSE6136) = GSE6136$ID_REF
GSE6136 = GSE6136[,-1]
cli = read.delim("C:/Users/benjk/Desktop/GSE6136_cli.txt", head = F)
cli = t(cli)
colnames(cli) = str_split_fixed(cli[1,],'_',2)[,2]
cli = cli[-1,]
cli = data.frame(cli)
cli[,'cell-type:ch1'] = str_split_fixed(cli$characteristics_ch1.1,': ',2)[,2]
cli[,'Ig clonality:ch1'] = str_split_fixed(cli$characteristics_ch1.3,': ',2)[,2]
colnames(cli)[colnames(cli) == 'cell-type:ch1'] = 'group'
cli$Row.names = cli[,1]

dat <- GSE6136
df <- log2(dat+1)
tmp <- names(table(cli$group))
sampleL <- split(cli[,1],f = cli$group)
test <- sd_selection(df,sampleL,0.01)
View(test)

}
\author{
Zhezhen Wang and Biniam Feleke
}
