% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_package.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test Data}
\description{
An idealised test data set, for demonstrating some of the NPS functions
}
\author{
Brendan Rocks \email{foss@brendanrocks.com}
}
\keyword{data}

