% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataSource}
\alias{dataSource}
\title{Description of NPS data}
\format{a \code{data frame} 

\describe{
\item{this}{this}
\item{this}{that}
}}
\usage{
dataSource
}
\description{
A dataset containing the original NPS data
}
\keyword{datasets}
