% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_paracoords.R
\name{myparacoords}
\alias{myparacoords}
\title{Parallel Coordinate Plot}
\usage{
myparacoords(dataSource)
}
\arguments{
\item{dataSource}{A data.frame with a column called "LCLUCI.labels"}
}
\description{
uses timelyportfolio's github function for a parallel coordinate plots
}
