% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps.R
\name{npc}
\alias{npc}
\title{Create Net Promoter Categories from Likelihood to Recommend Scores}
\usage{
npc(x, breaks = getOption("nps.breaks"))
}
\arguments{
\item{x}{A vector of \emph{Recommend} scores}

\item{breaks}{A \code{list} of length three, giving the integer Likert scale
points for \emph{Detractors}, \emph{Passives}, and \emph{Promoters},
respectively. The default is \code{list(0:6, 7:8, 9:10)}}
}
\value{
Net Promoter categories
}
\description{
This function produces Net Promoter Categories for \code{\link{numeric}} or
\code{\link{integer}} \emph{Recommend} data
}
\examples{
# The command below will generate Net Promoter categories for each point
# on a standard 0:10 Likelihood to Recommend scale
npc(0:10)

 # Here's how scores and categories map out. Notice that scores which are
 # 'off the scale' drop out as missing/invalid
data.frame(score = -2:12, category = npc(-2:12))

# When you have lots of data, summaries are useful
x <- sample(
  0:10, 1000, replace = TRUE,
  prob = c(0.02, 0.01, 0.01, 0.01, 0.01, 0.03, 0.03, 0.09, 0.22, 0.22, 0.35)
)

# A Histrogram of the Likelihood to Recommend scores we just generated
hist(x, breaks=-1:10)

# A look at the by nps category using summary
summary(npc(x))

# As above
table(npc(x))

# As a crosstabulation
table(x, npc(x))

nps(x)
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\seealso{
\code{\link{nps}}
}

