% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{NPS-package}
\alias{NPS}
\alias{NPS-package}
\title{Convenience Functions and Tests for Working With the Net Promoter Score (NPS)}
\description{
\tabular{ll}{
Package: \tab NPS\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-12-03\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Encoding: \tab UTF-8\cr
}
}
\details{
Small functions to make working with survey data in the context of a Net Promoter programme easier. Specifically, data transformation methods, some methods for examining the statistical properties of the NPS, such as its variance and standard errors, and some simple inferential testing procedures. Net Promoter and NPS are registered trademarks of Bain & Company, Satmetrix Systems and Fred Reichheld.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\keyword{package}

