\docType{package}
\name{NPS-package}
\alias{NPS}
\alias{NPS-package}
\title{Small convenience functions and tests for working with the Net Promoter Score (NPS)}
\description{
\tabular{ll}{ Package: \tab NPS\cr Type: \tab Package\cr
Version: \tab 0.1\cr Date: \tab 2013-01-29\cr License: \tab
GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\details{
This package provides a few small functions to make working
with survey data in an NPS context easier. Specifically,
data transformation methods, some methods for examining the
statistical properties of the NPS, such as its variance and
standard errors, and some simple inferential testing
procedures.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\keyword{package}

