% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps.R
\name{nps}
\alias{nps}
\title{Calculate a Net Promoter Score}
\usage{
nps(x, breaks = getOption("nps.breaks"), na.rm = FALSE)
}
\arguments{
\item{x}{A vector of \emph{Recommend} scores}

\item{breaks}{A \code{vector} of length three, giving integer Likert scale
break-points for Net Promoter categories. Best practice is to set the
values you'd like to use at the start of a session, with
\code{options(nps.breaks = c(0, 6, 8, 10))} (for the usual setting of 0-6
being \emph{Detractors}, 7-8 being \emph{Passives} and 9-10 being
\emph{Promoters}, also the package default), or whichever values you would
like. However, a vector may be supplied directly.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
a Net Promoter Score.
}
\description{
This function calculates a Net Promoter Score from a vector of
\emph{Recommend} scores, ideally \code{\link{numeric}} ones. An attempt will
be made to coerce \code{\link{factor}}, or \code{\link{character}} data.
\code{NA} values, either in the data, or generated by type coercion, are
automatically omitted from the calculation. No warning is given in the former
case. Net Promoter Scores generated are on a [-1,1] scale; you may want to
multiply them by 100 (and perhaps round them!) prior to presentation.
}
\examples{
# This will generate 1000 dummy Likelihood to Recommend reponses
x <- sample(
  0:10, 1000, replace = TRUE,
  prob = c(0.02, 0.01, 0.01, 0.01, 0.01, 0.03, 0.03, 0.09, 0.22, 0.22, 0.35)
)

# Here are the proportions of respondents giving each Likelihood to
# Recommend response
prop.table(table(x))

# Here's a histrogram of the scores
hist(
  x, breaks = -1:10,
  col = c(rep("red", 7), rep("yellow", 2), rep("green", 2))
)

# Here's a barplot. It's very similar, though for categorical responses
# it's often slightly easier to interpret.
barplot(
 prop.table(table(x)),
 col = c(rep("red", 7), rep("yellow", 2), rep("green", 2))
)

# Here's the NPS
nps(x)

# You can round it if you like
round(nps(x)) ; round(nps(x), 1)
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\seealso{
\code{\link{npc}}
}

