% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps.R
\name{npvar}
\alias{nps_var_}
\alias{npvar}
\title{Calculate the variance of a Net Promoter Score}
\usage{
nps_var_(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{\link{vector}} of length three, with numbers or proportions
of \emph{Detractors}, \emph{Passives}, and \emph{Promoters}, respectively}
}
\value{
\code{\link{numeric}}. The variance of the distribution, ranging from
  0 to 1.
}
\description{
This function calculates the Net Promoter Score variance, taking a
\code{\link{vector}} of length three, with numbers or proportions of
\emph{Detractors}, \emph{Passives}, and \emph{Promoters}, respectively.
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\seealso{
\code{\link{nps.var}}, a version which works on raw Recommend
  responses
}

