% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{sd_selection}
\alias{sd_selection}
\title{Selecting Highly Oscillating Transcripts}
\usage{
sd_selection(df, samplesL, cutoff = 0.01, method = "other")
}
\arguments{
\item{df}{A numeric matrix or data frame. The rows and columns represent
unique transcript IDs (geneID) and sample names, respectively.}

\item{samplesL}{A list of vectors, whose length is the number of states. Each
vector gives the sample names in a state, therefore most possibly to be a
vector of character or integer. Note that the vector s (sample names) has
to be among the column names of the R object 'df'.}

\item{cutoff}{A positive numeric value. Default is 0.01, if < 1,
automatically goes to select top x  transcripts using the a selecting
method (which is either the 'reference', 'other' stages or 'previous'
stage), e.g. by default it will select top 1\% of the transcripts.}

\item{method}{Selection of methods from 'reference', 'other', 'previous',
'itself' or 'longitudinal reference'.
\itemize{
\item 'reference': the reference (control) state is the first (or earliest) state.
\item 'previous': make sure sampleL is in the right order from benign to malign.
\item 'other': all other states in the dataset are gathered as a reference control state.
\item 'itself': make sure the cutoff is smaller than 1.
\item 'longitudinal reference': make sure control_df and control_samplesL are not NULL.
}}
}
\value{
\code{sd_selection} A list of data frames, whose length is the number
of states. The rows in each data frame are the filtered transcripts with
highest standard deviation selected from \code{df} and based on a cutoff
value assigned. Each resulting data frame represents a subset of the raw
input \code{df}, with the sample ID of the same state in the column.
}
\description{
\code{sd_selection} pre-selects highly oscillating transcripts
from the input dataset \code{df}. The dataset must contain multiple sample
groups (or 'states'). For each state, the function filters the dataset using
a cutoff value for standard deviation. The default cutoff value is 0.01
(i.e., higher than the top 1\% standard deviation).
}
\examples{
counts = matrix(sample(1:100,18),2,9)
colnames(counts) = 1:9
row.names(counts) = c('loci1','loci2')
cli = cbind(1:9,rep(c('state1','state2','state3'),each = 3))
colnames(cli) = c('samples','group')
samplesL <- split(cli[,1],f = cli[,'group'])
test_sd_selection <- sd_selection(counts, samplesL, 0.01)
}
\seealso{
\code\link{sd_selection.simulation}
}
\author{
Zhezhen Wang and Biniam Feleke
}
