% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{testdata}
\alias{testdata}
\title{This is data to be included in my package}
\description{
An idealised test data set, for demonstrating some of the NPS functions
}
\author{
Brendan Rocks \email{rocks.brendan@gmail.com}
}
\keyword{data}

